\name{komIgenNU-package}
\alias{komIgenNU-package}
\alias{komIgenNU}
\docType{package}
\title{
\packageTitle{komIgenNU}
}
\description{
\packageDescription{komIgenNU}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{komIgenNU}
\packageIndices{komIgenNU}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{komIgenNU}

Maintainer: \packageMaintainer{komIgenNU}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
